Broken file_input down into compoents
- Main file was edging up towards a thousand lines
- Currently broken down into, "events", "input" and "output"
