+++ src/mewbot/io/file_system/events.py
@@ -1,0 +1,171 @@
+import dataclasses
+
+import aiopath  # type: ignore
+
+from mewbot.api.v1 import InputEvent, OutputEvent
+
+
+@dataclasses.dataclass
+class FSInputEvent(InputEvent):
+    """
+    Base class for generic input events generated by the file system.
+    """
+
+
+@dataclasses.dataclass
+class FileFSInputEvent(FSInputEvent):
+    """
+    Base class for file related input events generated by the file system.
+    """
+
+    file_path: str
+    file_async_path: aiopath.AsyncPath
+
+
+@dataclasses.dataclass
+class CreatedFileFSInputEvent(FileFSInputEvent):
+    """
+    Called when a file is created - either at the location which is being monitored or inside a dir
+    which is being monitored.
+    """
+
+    # file_size: str
+    # entire_file: str
+    # file_read_mode: str
+
+
+@dataclasses.dataclass
+class UpdatedFileFSInputEvent(FileFSInputEvent):
+    """
+    Called when a file is updated - either at the location which is being monitored or inside a dir
+    which is being monitored.
+    """
+
+    # file_size: str
+    # entire_file: str
+    # file_read_mode: str
+
+
+@dataclasses.dataclass
+class MovedFileFSInputEvent(FileFSInputEvent):
+    """
+    Called when a file is updated - either at the location which is being monitored or inside a dir
+    which is being monitored.
+    """
+
+    # file_size: str
+    # entire_file: str
+    # file_read_mode: str
+
+
+@dataclasses.dataclass
+class DeletedFileFSInputEvent(FileFSInputEvent):
+    """
+    Called when a file is created - either at the location which is being monitored or inside a dir
+    which is being monitored.
+    Cannot provide any more information without caching a lot of details about the FS.
+    """
+
+
+@dataclasses.dataclass
+class DirFSInputEvent(FSInputEvent):
+    """
+    Base class for file related input events generated by the file system.
+    """
+
+    dir_path: str
+    dir_async_path: aiopath.AsyncPath
+
+
+@dataclasses.dataclass
+class CreatedDirFSInputEvent(DirFSInputEvent):
+    """
+    A dir has been created - either in the dir we're monitoring - if we're monitoring one.
+    Or at the given monitor location.
+    """
+
+
+@dataclasses.dataclass
+class UpdatedDirFSInputEvent(DirFSInputEvent):
+    """
+    A dir has been updated - either in the dir we're monitoring - if we're monitoring one.
+    What this mostly means is the name of the dir has been updated.
+    If the contents have been updated, it'll spew events of the appropriate type.
+    """
+
+
+@dataclasses.dataclass
+class MovedDirFSInputEvent(DirFSInputEvent):
+    """
+    A dir has been updated - either in the dir we're monitoring - if we're monitoring one.
+    What this mostly means is the name of the dir has been updated.
+    If the contents have been updated, it'll spew events of the appropriate type.
+    """
+
+    dir_src_path: str
+    dir_src_async_path: aiopath.AsyncPath
+    # dir_path and dir_async_path are where the dir CURRENTLY IS
+
+
+@dataclasses.dataclass
+class DeletedDirFSInputEvent(DirFSInputEvent):
+    """
+    A dir has been deleted - either in the dir we're monitoring - if we're monitoring one.
+    """
+
+
+@dataclasses.dataclass
+class FSOutputEvent(OutputEvent):
+    """
+    Base class for generic input events generated by the file system.
+    """
+
+
+@dataclasses.dataclass
+class FileFSOutputEvent(OutputEvent):
+    """
+    Base class for generic input events generated by the file system.
+    """
+
+
+@dataclasses.dataclass
+class CreateFileFSOutputEvent(FileFSOutputEvent):
+    """
+    Creates a file with the given name.
+    For security (and sanity) currently only created files in the designated output folder.
+    """
+
+    file_name: str
+    file_contents: str
+
+
+@dataclasses.dataclass
+class AppendFileFSOutputEvent(FileFSOutputEvent):
+    """
+    Creates a file with the given name.
+    For security (and sanity) currently only created files in the designated output folder.
+    """
+
+    file_name: str
+    file_contents: str
+
+
+@dataclasses.dataclass
+class OverwriteFileFSOutputEvent(FileFSOutputEvent):
+    """
+    Creates a file with the given name.
+    For security (and sanity) currently only created files in the designated output folder.
+    """
+
+    file_name: str
+    file_contents: str
+
+
+@dataclasses.dataclass
+class DeleteFileFSOutputEvent(FileFSOutputEvent):
+    """
+    Creates a file with the given name.
+    For security (and sanity) currently only created files in the designated output folder.
+    """
+
+    file_name: str
