+++ src/mewbot/io/file_system/input.py
@@ -3,13 +3,11 @@
 from __future__ import annotations
 
 import os.path
-from typing import Optional, Set, Sequence, Type, Union, Any
+from typing import Optional, Set, Type, Any
 
 import asyncio
-import dataclasses
 import datetime
 import logging
-import pathlib
 
 import aiopath  # type: ignore
 import watchfiles
@@ -17,241 +15,18 @@ import watchdog  # type: ignore
 from watchdog.events import FileSystemEvent, FileSystemEventHandler  # type: ignore
 from watchdog.observers import Observer  # type: ignore
 
-from mewbot.api.v1 import IOConfig, Input, Output, InputEvent, OutputEvent
-
-
-@dataclasses.dataclass
-class FSInputEvent(InputEvent):
-    """
-    Base class for generic input events generated by the file system.
-    """
-
-
-@dataclasses.dataclass
-class FileFSInputEvent(FSInputEvent):
-    """
-    Base class for file related input events generated by the file system.
-    """
-
-    file_path: str
-    file_async_path: aiopath.AsyncPath
-
-
-@dataclasses.dataclass
-class CreatedFileFSInputEvent(FileFSInputEvent):
-    """
-    Called when a file is created - either at the location which is being monitored or inside a dir
-    which is being monitored.
-    """
-
-    # file_size: str
-    # entire_file: str
-    # file_read_mode: str
-
-
-@dataclasses.dataclass
-class UpdatedFileFSInputEvent(FileFSInputEvent):
-    """
-    Called when a file is updated - either at the location which is being monitored or inside a dir
-    which is being monitored.
-    """
-
-    # file_size: str
-    # entire_file: str
-    # file_read_mode: str
-
-
-@dataclasses.dataclass
-class MovedFileFSInputEvent(FileFSInputEvent):
-    """
-    Called when a file is updated - either at the location which is being monitored or inside a dir
-    which is being monitored.
-    """
-
-    # file_size: str
-    # entire_file: str
-    # file_read_mode: str
-
-
-@dataclasses.dataclass
-class DeletedFileFSInputEvent(FileFSInputEvent):
-    """
-    Called when a file is created - either at the location which is being monitored or inside a dir
-    which is being monitored.
-    Cannot provide any more information without caching a lot of details about the FS.
-    """
-
-
-@dataclasses.dataclass
-class DirFSInputEvent(FSInputEvent):
-    """
-    Base class for file related input events generated by the file system.
-    """
-
-    dir_path: str
-    dir_async_path: aiopath.AsyncPath
-
-
-@dataclasses.dataclass
-class CreatedDirFSInputEvent(DirFSInputEvent):
-    """
-    A dir has been created - either in the dir we're monitoring - if we're monitoring one.
-    Or at the given monitor location.
-    """
-
-
-@dataclasses.dataclass
-class UpdatedDirFSInputEvent(DirFSInputEvent):
-    """
-    A dir has been updated - either in the dir we're monitoring - if we're monitoring one.
-    What this mostly means is the name of the dir has been updated.
-    If the contents have been updated, it'll spew events of the appropriate type.
-    """
-
-
-@dataclasses.dataclass
-class MovedDirFSInputEvent(DirFSInputEvent):
-    """
-    A dir has been updated - either in the dir we're monitoring - if we're monitoring one.
-    What this mostly means is the name of the dir has been updated.
-    If the contents have been updated, it'll spew events of the appropriate type.
-    """
-
-    dir_src_path: str
-    dir_src_async_path: aiopath.AsyncPath
-    # dir_path and dir_async_path are where the dir CURRENTLY IS
-
-
-@dataclasses.dataclass
-class DeletedDirFSInputEvent(DirFSInputEvent):
-    """
-    A dir has been deleted - either in the dir we're monitoring - if we're monitoring one.
-    """
-
-
-@dataclasses.dataclass
-class FSOutputEvent(OutputEvent):
-    """
-    Base class for generic input events generated by the file system.
-    """
-
-
-@dataclasses.dataclass
-class FileFSOutputEvent(OutputEvent):
-    """
-    Base class for generic input events generated by the file system.
-    """
-
-
-@dataclasses.dataclass
-class CreateFileFSOutputEvent(FileFSOutputEvent):
-    """
-    Creates a file with the given name.
-    For security (and sanity) currently only created files in the designated output folder.
-    """
-
-    file_name: str
-    file_contents: str
-
-
-@dataclasses.dataclass
-class AppendFileFSOutputEvent(FileFSOutputEvent):
-    """
-    Creates a file with the given name.
-    For security (and sanity) currently only created files in the designated output folder.
-    """
-
-    file_name: str
-    file_contents: str
-
-
-@dataclasses.dataclass
-class OverwriteFileFSOutputEvent(FileFSOutputEvent):
-    """
-    Creates a file with the given name.
-    For security (and sanity) currently only created files in the designated output folder.
-    """
-
-    file_name: str
-    file_contents: str
-
-
-@dataclasses.dataclass
-class DeleteFileFSOutputEvent(FileFSOutputEvent):
-    """
-    Creates a file with the given name.
-    For security (and sanity) currently only created files in the designated output folder.
-    """
-
-    file_name: str
-
-
-class FileSystemIO(IOConfig):
-
-    _input: Optional[Union[FileTypeFSInput, DirTypeFSInput]] = None
-    _output: Optional[FileSystemOutput] = None
-
-    _input_path: Optional[str] = None
-    _input_path_type: str = "not_set"
-    _output_path: Optional[str] = None
-
-    @property
-    def input_path(self) -> Optional[str]:
-        return self._input_path
-
-    @input_path.setter
-    def input_path(self, input_path: str) -> None:
-        self._input_path = input_path
-
-    @property
-    def input_path_type(self) -> str:
-        """
-        When starting this class you need to set the type of resource you are monitoring.
-        This is due to limitations of the underlying libraries used to do the actual monitoring.
-        """
-        return self._input_path_type
-
-    @input_path_type.setter
-    def input_path_type(self, input_path_type: str) -> None:
-        assert input_path_type in (
-            "dir",
-            "file",
-        ), f"input_path_type couldn't be set as {input_path_type}"
-        self._input_path_type = input_path_type
-
-    @property
-    def output_path(self) -> Optional[str]:
-        return self._output_path
-
-    @output_path.setter
-    def output_path(self, output_path: str) -> None:
-        self._output_path = output_path
-
-    def get_inputs(self) -> Sequence[Input]:
-
-        assert self._input_path_type in ("dir", "file",), (
-            f"input_path_type must be properly set before startup - "
-            f"{self._input_path_type} is not proper"
-        )
-
-        if not self._input:
-
-            if self._input_path_type == "file":
-                self._input = FileTypeFSInput(self._input_path)
-            elif self._input_path_type == "dir":
-                self._input = DirTypeFSInput(self._input_path)
-            else:
-                raise NotImplementedError(
-                    f"{self._input_path_type} not good. Options are 'dir' and 'file'"
-                )
-
-        return [self._input]
-
-    def get_outputs(self) -> Sequence[Output]:
-        if not self._output:
-            self._output = FileSystemOutput(self._output_path)
-
-        return [self._output]
+from mewbot.api.v1 import Input, InputEvent
+from mewbot.io.file_system.events import (
+    FSInputEvent,
+    CreatedFileFSInputEvent,
+    UpdatedFileFSInputEvent,
+    MovedFileFSInputEvent,
+    DeletedFileFSInputEvent,
+    CreatedDirFSInputEvent,
+    UpdatedDirFSInputEvent,
+    MovedDirFSInputEvent,
+    DeletedDirFSInputEvent,
+)
 
 
 class FileTypeFSInput(Input):
@@ -265,8 +40,10 @@ class FileTypeFSInput(Input):
     """
 
     _logger: logging.Logger
+
     _input_path: Optional[str] = None  # A location on the file system to monitor
     _input_path_exists: bool = False
+    _input_path_type: Optional[str] = None
 
     def __init__(self, input_path: Optional[str] = None) -> None:
         super(Input, self).__init__()  # pylint: disable=bad-super-call
@@ -276,13 +53,6 @@ class FileTypeFSInput(Input):
         self._logger = logging.getLogger(__name__ + ":" + type(self).__name__)
 
         if input_path is None or not os.path.exists(input_path):
-
-            self.watcher = None
-            self._input_path_exists = False
-            self._input_path_type = None
-
-        # Cannot await in this context - this is being done before loop start
-        elif not os.path.exists(input_path):
 
             self.watcher = None
             self._input_path_exists = False
@@ -835,10 +605,38 @@ class DirTypeFSInput(Input):
         print(event)
         print(type(event))
 
-        # Not ... entirely sure why this is happening - will work on it
+        # Filter null events
         if event is None:
             return
 
+        if isinstance(
+            event,
+            (
+                watchdog.events.FileCreatedEvent,
+                watchdog.events.FileModifiedEvent,
+                watchdog.events.FileMovedEvent,
+                watchdog.events.FileSystemMovedEvent,
+                watchdog.events.FileDeletedEvent,
+            ),
+        ):
+            await self._process_file_event(event)
+        elif isinstance(
+            event,
+            (
+                watchdog.events.DirCreatedEvent,
+                watchdog.events.DirModifiedEvent,
+                watchdog.events.DirMovedEvent,
+                watchdog.events.DirDeletedEvent,
+            ),
+        ):
+            await self._process_dir_event(event)
+        else:
+            self._logger.info("Unhandled event in _process_event - %s", event)
+
+    async def _process_file_event(self, event: FileSystemEvent) -> None:
+        """
+        Take an event and process it before putting it on the wire.
+        """
         # FILES
         if isinstance(event, watchdog.events.FileCreatedEvent):
             print("Is this going on the wire?")
@@ -882,8 +680,13 @@ class DirTypeFSInput(Input):
                 )
             )
 
+    async def _process_dir_event(self, event: FileSystemEvent) -> None:
+        """
+        Take an event and process it before putting it on the wire.
+        """
+
         # DIRS
-        elif isinstance(event, watchdog.events.DirCreatedEvent):
+        if isinstance(event, watchdog.events.DirCreatedEvent):
             await self.send(
                 CreatedDirFSInputEvent(
                     dir_path=event.src_path,
@@ -916,20 +719,16 @@ class DirTypeFSInput(Input):
                 )
             )
 
-        else:
-
-            self._logger.info("Unhandled event in _process_event - %s", event)
-
-    def watch(
-        self
-    ) -> None:
+    def watch(self) -> None:
         """
         Use watchdog in a seperate thread to watch a dir for changes.
         """
         handler = _EventHandler(queue=self._internal_queue, loop=asyncio.get_event_loop())
 
         self._watchdog_observer = Observer()
-        self._watchdog_observer.schedule(event_handler=handler, path=self._input_path, recursive=True)
+        self._watchdog_observer.schedule(
+            event_handler=handler, path=self._input_path, recursive=True
+        )
         self._watchdog_observer.start()
 
         self._logger.info("Started _watchdog_observer")
@@ -937,39 +736,6 @@ class DirTypeFSInput(Input):
         self._watchdog_observer.join(10)
 
         asyncio.get_event_loop().call_soon_threadsafe(self._internal_queue.put_nowait, None)
-
-
-class FileSystemOutput(Output):
-
-    _logger: logging.Logger
-    _output_path: Optional[str]  # A location on the file system to monitor
-
-    def __init__(self, output_path: Optional[str] = None) -> None:
-        super(Output, self).__init__()  # pylint: disable=bad-super-call
-
-        self._output_path = output_path
-        self._logger = logging.getLogger(__name__ + "FileSystemOutput")
-
-    @staticmethod
-    def consumes_outputs() -> Set[Type[OutputEvent]]:
-        """
-        Defines the set of output events that this Output class can consume
-        :return:
-        """
-        return {
-            CreateFileFSOutputEvent,
-            AppendFileFSOutputEvent,
-            OverwriteFileFSOutputEvent,
-            DeleteFileFSOutputEvent,
-        }
-
-    async def output(self, event: OutputEvent) -> bool:
-        """
-        Does the work of transmitting the event to the world.
-        :param event:
-        :return:
-        """
-        raise NotImplementedError("Not yet ready for prime time")
 
 
 class _EventHandler(FileSystemEventHandler):  # type: ignore
